fix: multi-line macros now handled correctly
There was a problem during the processing of preprocessor macros that spanned more than one line. In these cases, no annotation could be derived and the analysis assumed that these should be `endif` lines which was incorrect. I changed the analysis logic to now annotate all lines based on the annotation to which they belong instead of tracking artifact lines as well. Lines that are not part of a mapping are annotated with the root mapping (i.e., True)